examples/image-processing/requirements.txt
click
pillow 
pyzmq 
